backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.35
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2

# SSH
asyncssh==2.17.0

# Celery + Redis
celery==5.4.0
redis==5.1.1
redbeat==2.2.0
kombu==5.3.4

# InfluxDB
influxdb-client==1.44.0

# Auth / Email
PyJWT==2.9.0
aiosmtplib==3.0.2
email-validator==2.2.0
passlib[bcrypt]==1.7.4

# Twilio
twilio==9.3.3

# HTTP
httpx==0.27.2

# Utils
python-multipart==0.0.12
python-dotenv==1.0.1
